created notifications dropdown menu to conditionally render based on user type
