+++ client/src/components/Header/Header.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { AiOutlineMenu } from 'react-icons/ai';
-import { IoNotificationsSharp } from 'react-icons/io5';
 import { CgProfile } from 'react-icons/cg';
 import {
   Flex,
@@ -15,16 +14,15 @@ import {
   MenuDivider,
 } from '@chakra-ui/react';
 import Navigation from './Navigation.jsx';
+import Notifications from './Notifications.jsx';
 
 export default function Header() {
-  const notifications = ['interview in one hour'];
-  // const user = {
-  //   firstName: 'Donna',
-  //   lastName: 'Wong',
-  // };
+  const user = {
+    type: 'employer',
+    firstName: 'Donna',
+    lastName: 'Wong',
+  };
 
-  const user = undefined;
-  const userState = 'guest';
   const firstName = user ? user.firstName : 'Guest';
   const salutation = `Welcome, ${firstName}!`;
 
@@ -39,18 +37,9 @@ export default function Header() {
           <MenuButton as={Button}>
             <AiOutlineMenu />
           </MenuButton>
-          <Navigation userState={userState} />
+          <Navigation user={user} />
         </Menu>
-        <Menu>
-          <MenuButton as={Button}>
-            <IoNotificationsSharp />
-          </MenuButton>
-          <MenuList>
-            { notifications.length === 0 ? <MenuItem>No New Notifications</MenuItem> : notifications.map((notification) => (
-              <MenuItem key={notification}>{ notification }</MenuItem>
-            ))}
-          </MenuList>
-        </Menu>
+        { user.type !== 'guest' ? <Notifications /> : null }
         <Menu>
           <MenuButton as={Button}>
             <CgProfile />
