+++ client/src/components/Header/Notifications.jsx
@@ -1,12 +1,31 @@
 // LIBRARY IMPORTS
 import React from 'react';
+import {
+  Menu,
+  MenuList,
+  MenuItem,
+  MenuButton,
+  Button,
+} from '@chakra-ui/react';
+import { IoNotificationsSharp } from 'react-icons/io5';
 
 // LOCAL IMPORTS
 
 // COMPONENT
 function Notifications() {
+  const notifications = ['interview in one hour'];
+
   return (
-    <div>Notifications</div>
+    <Menu>
+      <MenuButton as={Button}>
+        <IoNotificationsSharp />
+      </MenuButton>
+      <MenuList>
+        { notifications.length === 0 ? <MenuItem>No New Notifications</MenuItem> : notifications.map((notification) => (
+          <MenuItem key={notification}>{ notification }</MenuItem>
+        ))}
+      </MenuList>
+    </Menu>
   );
 }
